+++ src/placementEngine.js
@@ -70,27 +70,26 @@ function incrementCounter (keyPath, ts) {
 }
 
 /**
- * Creates a new placement context that contains functions and objects to be
+ * Creates a new placement context that contains objects and functions to be
  * made available to the constraint queries.
  *
  * @private
  */
-function createContext (window, user) {
-  const uaParser = new UAParser(get('navigator.userAgent', window))
+function createContext (userAgent, user) {
+  const uaParser = new UAParser(userAgent)
 
   return {
+    // Objects
+    browser: uaParser.getBrowser(),
+    device: uaParser.getDevice(),
+    os: uaParser.getOS(),
     user,
     window,
 
-    // Utility functions
+    // Functions
     age,
     lower: toLower,
-    upper: toUpper,
-
-    // User agent obejcts
-    browser: uaParser.getBrowser(),
-    device: uaParser.getDevice(),
-    os: uaParser.getOS()
+    upper: toUpper
   }
 }
 
@@ -103,13 +102,14 @@ function createContext (window, user) {
  * The placement engine returns a promise that resolves to an object containing
  * the placed promos and the callback functions.
  *
- * @param {Window} window The window object.
+ * @param {Storage} storage The storage object.
+ * @param {String} userAgent The user agent string.
  * @param {Array} promos The list of the candidate promos.
  * @returns {Promise} A promise.
  */
-export default function placementEngine (window, promos) {
-  const updateUser = (f, user) => set(window.localStorage, f(user))
-  let user = updateUser(incrementVisits, get(window.localStorage))
+export default function placementEngine ({ promos, storage, userAgent }) {
+  const updateUser = (f, user) => set(storage, f(user))
+  let user = updateUser(incrementVisits, get(storage))
 
   const onClick = promo => {
     const f = trackEngagement(promo, timestamp())
@@ -126,7 +126,8 @@ export default function placementEngine (window, promos) {
     updateUser(f, user)
   }
 
-  const context = createContext(window, user)
+  // Place the promos.
+  const context = createContext(userAgent, user)
   const placedPromos = placePromos(context, promos)
 
   // Return a promise containing the placed promos, and the callback functions.