+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Return user from userState `set` function
 * Refactor the placement engine to use promises
-* Swap arguments to `placementEngine` function
+* Update `placementEngine` function to take `promos`, `storage`, and `userAgent` arguments
 
 ## 3.0.0 (2019-04-30)